chore(action.yml): uncomment and configure scan_ansible_playbook job
The scan_ansible_playbook job was previously commented out. This commit uncommented the job and configured it to scan the Ansible playbook. The job runs on the ubuntu-latest environment and consists of several steps.

The steps include:
- Checking out the repository using the actions/checkout@v4 action.
- Creating a folder for storing the scan results.
- Running the kics Scan using the checkmarx/kics-github-action@v1.7.0 action. The scan is performed on the current directory and the results are saved in the specified output_path in JSON and SARIF formats. The fail_on parameter determines the severity level at which the scan should fail. The enable_jobs_summary parameter is set to true to enable the display of job summary.
- Uploading the SARIF file using the github/codeql-action/upload-sarif@v2 action. The SARIF file is located in the results directory and the wait-for-processing parameter is set to true to wait for the SARIF file to be processed.

Additionally, the run_ansible_playbook job was modified to include a dependency on the scan_ansible_playbook job. This ensures that the scan job is executed before the playbook is run.