chore(azure-pipelines.yml): update pipeline steps for security scan artifact handling and medium issue check
The pipeline steps for handling the security scan artifact result have been updated. Instead of using a Bash script to query the scan artifact result, the artifact is now downloaded using the "DownloadPipelineArtifact" task.

A new task "Check for MEDIUM issues in the scan result" has been added. This task uses a Bash script to check for MEDIUM issues in the scan result. It retrieves the path to the sarif file and uses jq to extract the number of MEDIUM issues. If there are more than 1 MEDIUM issue, the pipeline fails. Otherwise, the pipeline continues.

These changes improve the readability and maintainability of the pipeline and provide better handling of the security scan artifact result.