fix(azure-pipelines.yml): update script to check for more than 1 'warning' level issue in SARIF file and fail pipeline if found
The script in the azure-pipelines.yml file has been updated to check for more than 1 'warning' level issue in the SARIF file. If more than 1 issue is found, the pipeline will fail. This change improves the pipeline's quality control by ensuring that only a limited number of 'warning' level issues are allowed.