chore(azure-pipelines.yml): add conditional check for number of issues found in the code analysis results to fail the pipeline if more than one issue is found
The code in the azure-pipelines.yml file has been modified to include a conditional check for the number of issues found in the code analysis results. If the number of issues is greater than one, the pipeline will fail and display a message indicating the number of issues found. If no issues are found, the pipeline will continue. This change improves the pipeline's ability to detect and handle code issues during the build process.