fix(azure-pipelines.yml): fix syntax error in if-else statement
The if-else statement in the pipeline script had a syntax error. The "elseif" keyword was used instead of "fi" to close the previous if statement. This caused the pipeline to fail when there were no issues found. The fix corrects the syntax by replacing "elseif" with "fi" to properly close the if statement.