chore(azure-pipelines.yml): add support for generating a summary of Ansible playbook output
The changes in this commit add support for generating a summary of the Ansible playbook output in the Azure Pipelines YAML file. This is done by redirecting the output of the `ansible-playbook` command to a log file and then creating a summary markdown file that includes the contents of the log file. The summary markdown file is then uploaded as a build artifact. This change improves the visibility and accessibility of the Ansible playbook output during the pipeline execution.